fix(session): Fix invalid headers
MsgType<35> was not guaranteed to be third, and we were relying on the
user to provide SendingTime.
Since the goal of this library is to
provide simple and fast scripting, we now define the sending time on
message creation. This is correct in terms of `msg_to_fix` creating a
valid FIX message and `msg_to_fix` is used within `send` so will still
be accurate upon sending.